+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-0.3.0.0
+0.2.1.0
 =======
 
   * Add a `shouldPerform` expectation